05_Authentication && 1:N model relation

+++ posts/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render, redirect, get_list_or_404, get_object_or_404
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST
 from .models import Post, Image
 from .forms import PostForm, ImageForm
 
@@ -9,12 +11,15 @@ def list(request):  # index임
     context = {'posts':posts,}
     return render(request, 'posts/list.html', context)
 
+@require_POST
+@login_required
 def create(request):
     if request.method == 'POST':
         post_form = PostForm(request.POST)
         if post_form.is_valid():
-            post = post_form.save() # 게시글 내용 처리 끝
-            
+            post = post_form.save(commit=False) # 게시글 내용 처리 끝
+            post.user = request.user
+            post.save()
             # 이미지 여러장 받기
             for image in request.FILES.getlist('file'):
                 request.FILES['file'] = image
@@ -41,8 +46,13 @@ def create(request):
 #     }
 #     return render(request, 'posts/read.html', post.pk)
 
+@login_required
 def update(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
+    
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     if request.method == 'POST':
         post_form = PostForm(request.POST, instance=post) 
         if post_form.is_valid():
@@ -50,17 +60,23 @@ def update(request, post_pk):
             return redirect('posts:list')
     else:
         post_form = PostForm(instance=post)
+    
     context = {
         'post_form': post_form,
+        'post': post,
     }
     return render(request, 'posts/form.html', context)
 
+@login_required
 def delete(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     if request.method == 'POST':
         post.delete()
         return redirect('posts:list')
     else:
-        return redirect('posts:read', post_pk)
+        return redirect('posts:list')
